requirements.txt
openai-whisper
pyannote.audio
python-dotenv
torch
torchaudio
soundfile
ffmpeg-python
numpy
scipy
librosa